Remove Favorite and UnFavorite buttons and add <a> tags

+++ MobileWorld/Controllers/UserController.cs
@@ -37,16 +37,25 @@ namespace MobileWorld.Controllers
         //}
         public IActionResult Favorite(string adId, string userId)
         {
+            if(userId == null || adId==null)
+            {
+                return Redirect("~/Identity/Account/Login");
+            }
 
             bool result = this.userService.AddToFavorites(adId, userId);
             return Ok();
         }
-
         public IActionResult UnFavorite(string adId, string userId)
         {
+            if(adId == null || adId == null)
+            {
+                return Redirect("~/Identity/Account/Login");
+            }
             bool result = this.userService.RemoveFromFavorites(adId, userId);
             return Ok();
         }
+
+        
 
     }
 }
